social-network-hw3/post-api-gateway/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
protobuf
pytest
httpx
pytest-asyncio